#1311 clean up tests
