+++ tests/unit/test_spatial_methods/test_finite_volume/test_finite_volume.py
@@ -1,14 +1,12 @@
 #
-# Test for the operator class
+# Tests for the Finite Volume Method
 #
 import pybamm
 from tests import (
     get_mesh_for_testing,
     get_p2d_mesh_for_testing,
     get_1p1d_mesh_for_testing,
-    get_cylindrical_mesh_for_testing,
 )
-
 import numpy as np
 from scipy.sparse import kron, eye
 import unittest
@@ -16,18 +14,23 @@ import unittest
 
 class TestFiniteVolume(unittest.TestCase):
     def test_node_to_edge_to_node(self):
-        # create discretisation
+        # Create discretisation
         mesh = get_mesh_for_testing()
         fin_vol = pybamm.FiniteVolume()
         fin_vol.build(mesh)
         n = mesh["negative electrode"].npts
 
         # node to edge
-        c = pybamm.Vector(np.ones(n), domain=["negative electrode"])
+        c = pybamm.StateVector(slice(0, n), domain=["negative electrode"])
+        y_test = np.ones(n)
         diffusivity_c_ari = fin_vol.node_to_edge(c, method="arithmetic")
-        np.testing.assert_array_equal(diffusivity_c_ari.evaluate(), np.ones((n + 1, 1)))
+        np.testing.assert_array_equal(
+            diffusivity_c_ari.evaluate(None, y_test), np.ones((n + 1, 1))
+        )
         diffusivity_c_har = fin_vol.node_to_edge(c, method="harmonic")
-        np.testing.assert_array_equal(diffusivity_c_har.evaluate(), np.ones((n + 1, 1)))
+        np.testing.assert_array_equal(
+            diffusivity_c_har.evaluate(None, y_test), np.ones((n + 1, 1))
+        )
 
         # edge to node
         d = pybamm.StateVector(slice(0, n + 1), domain=["negative electrode"])
@@ -73,14 +76,11 @@ class TestFiniteVolume(unittest.TestCase):
             fin_vol.concatenation(edges)
 
     def test_discretise_diffusivity_times_spatial_operator(self):
-        # Set up
-        whole_cell = ["negative electrode", "separator", "positive electrode"]
-
-        # create discretisation
+        # Setup mesh and discretisation
         mesh = get_mesh_for_testing()
         spatial_methods = {"macroscale": pybamm.FiniteVolume()}
         disc = pybamm.Discretisation(mesh, spatial_methods)
-
+        whole_cell = ["negative electrode", "separator", "positive electrode"]
         combined_submesh = mesh.combine_submeshes(*whole_cell)
 
         # Discretise some equations where averaging is needed
@@ -104,7 +104,8 @@ class TestFiniteVolume(unittest.TestCase):
             -2 * pybamm.div(var * pybamm.grad(var) + 2 * pybamm.grad(var)),
             pybamm.laplacian(var),
         ]:
-            # Check that the equation can be evaluated in each case
+            # Check that the equation can be evaluated for different combinations
+            # of boundary conditions
             # Dirichlet
             disc.bcs = {
                 var.id: {
@@ -141,1186 +142,8 @@ class TestFiniteVolume(unittest.TestCase):
             eqn_disc = disc.process_symbol(eqn)
             eqn_disc.evaluate(None, y_test)
 
-    def test_grad_1plus1d(self):
-        mesh = get_1p1d_mesh_for_testing()
-        spatial_methods = {"macroscale": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        a = pybamm.Variable(
-            "a",
-            domain=["negative electrode"],
-            auxiliary_domains={"secondary": "current collector"},
-        )
-        b = pybamm.Variable(
-            "b",
-            domain=["separator"],
-            auxiliary_domains={"secondary": "current collector"},
-        )
-        c = pybamm.Variable(
-            "c",
-            domain=["positive electrode"],
-            auxiliary_domains={"secondary": "current collector"},
-        )
-        var = pybamm.concatenation(a, b, c)
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Vector(np.linspace(0, 1, 15)), "Neumann"),
-                "right": (pybamm.Vector(np.linspace(0, 1, 15)), "Neumann"),
-            }
-        }
-
-        disc.bcs = boundary_conditions
-        disc.set_variable_slices([var])
-        grad_eqn_disc = disc.process_symbol(pybamm.grad(var))
-
-        # Evaulate
-        combined_submesh = mesh.combine_submeshes(*var.domain)
-        linear_y = np.outer(np.linspace(0, 1, 15), combined_submesh.nodes).reshape(
-            -1, 1
-        )
-        expected = np.outer(
-            np.linspace(0, 1, 15), np.ones_like(combined_submesh.edges)
-        ).reshape(-1, 1)
-        np.testing.assert_array_almost_equal(
-            grad_eqn_disc.evaluate(None, linear_y), expected
-        )
-
-    def test_grad_div_shapes_Dirichlet_bcs(self):
-        """
-        Test grad and div with Dirichlet boundary conditions (applied by grad on var)
-        in Cartesian coordinates
-        """
-        whole_cell = ["negative electrode", "separator", "positive electrode"]
-        # create discretisation
-        mesh = get_mesh_for_testing()
-        spatial_methods = {"macroscale": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-        combined_submesh = mesh.combine_submeshes(*whole_cell)
-
-        # grad(1) = 0
-        var = pybamm.Variable("var", domain=whole_cell)
-        grad_eqn = pybamm.grad(var)
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(1), "Dirichlet"),
-                "right": (pybamm.Scalar(1), "Dirichlet"),
-            }
-        }
-        disc.bcs = boundary_conditions
-        disc.set_variable_slices([var])
-        grad_eqn_disc = disc.process_symbol(grad_eqn)
-        constant_y = np.ones_like(combined_submesh.nodes[:, np.newaxis])
-        np.testing.assert_array_equal(
-            grad_eqn_disc.evaluate(None, constant_y),
-            np.zeros_like(combined_submesh.edges[:, np.newaxis]),
-        )
-        # div(grad(x))
-        linear_y = combined_submesh.nodes
-        N = pybamm.grad(var)
-        div_eqn = pybamm.div(N)
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(0), "Dirichlet"),
-                "right": (pybamm.Scalar(1), "Dirichlet"),
-            }
-        }
-        disc.bcs = boundary_conditions
-        # grad(x) = 1
-        grad_eqn_disc = disc.process_symbol(grad_eqn)
-        np.testing.assert_array_almost_equal(
-            grad_eqn_disc.evaluate(None, linear_y),
-            np.ones_like(combined_submesh.edges[:, np.newaxis]),
-        )
-        # div(grad(x)) = 0
-        div_eqn_disc = disc.process_symbol(div_eqn)
-        np.testing.assert_array_almost_equal(
-            div_eqn_disc.evaluate(None, linear_y),
-            np.zeros_like(combined_submesh.nodes[:, np.newaxis]),
-        )
-
-    def test_cylindrical_grad_div_shapes_Dirichlet_bcs(self):
-        """
-        Test grad and div with Dirichlet boundary conditions (applied by grad on var)
-        in cylindrical polar coordinates
-        """
-        # create discretisation
-        mesh = get_cylindrical_mesh_for_testing()
-        spatial_methods = {"current collector": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        submesh = mesh["current collector"]
-        npts = submesh.npts
-        npts_edges = submesh.npts + 1
-
-        # grad
-        var = pybamm.Variable(
-            "var",
-            domain=["current collector"],
-        )
-        grad_eqn = pybamm.grad(var)
-        # grad(1) = 0
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(1), "Dirichlet"),
-                "right": (pybamm.Scalar(1), "Dirichlet"),
-            }
-        }
-        disc.bcs = boundary_conditions
-        disc.set_variable_slices([var])
-        grad_eqn_disc = disc.process_symbol(grad_eqn)
-        constant_y = np.ones((npts, 1))
-        np.testing.assert_array_equal(
-            grad_eqn_disc.evaluate(None, constant_y), np.zeros((npts_edges, 1))
-        )
-        # grad(r) == 1
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(submesh.edges[0]), "Dirichlet"),
-                "right": (pybamm.Scalar(1), "Dirichlet"),
-            }
-        }
-        disc.bcs = boundary_conditions
-        y_linear = submesh.nodes
-        grad_eqn_disc = disc.process_symbol(grad_eqn)
-        np.testing.assert_array_almost_equal(
-            grad_eqn_disc.evaluate(None, y_linear), np.ones((npts_edges, 1))
-        )
-        # div(grad r^2) = 4
-        N = pybamm.grad(var)
-        div_eqn = pybamm.div(N)
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(submesh.edges[0] ** 2), "Dirichlet"),
-                "right": (pybamm.Scalar(1), "Dirichlet"),
-            }
-        }
-        disc.bcs = boundary_conditions
-        div_eqn_disc = disc.process_symbol(div_eqn)
-        y_squared = submesh.nodes ** 2
-        div_eval = div_eqn_disc.evaluate(None, y_squared)
-        np.testing.assert_array_almost_equal(
-            div_eval[1:-1],
-            4 * np.ones((npts - 2, 1)),
-        )
-
-    def test_spherical_grad_div_shapes_Dirichlet_bcs(self):
-        """
-        Test grad and div with Dirichlet boundary conditions (applied by grad on var)
-        in spherical polar coordinates
-        """
-        # create discretisation
-        mesh = get_1p1d_mesh_for_testing()
-        spatial_methods = {"negative particle": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-        submesh = mesh["negative particle"]
-        npts = submesh.npts
-        sec_npts = mesh["negative electrode"].npts * mesh["current collector"].npts
-        total_npts = npts * sec_npts
-        total_npts_edges = (npts + 1) * sec_npts
-
-        # grad
-        var = pybamm.Variable(
-            "var",
-            domain=["negative particle"],
-            auxiliary_domains={
-                "secondary": "negative electrode",
-                "tertiary": "current collector",
-            },
-        )
-        grad_eqn = pybamm.grad(var)
-        # grad(1) = 0
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(1), "Dirichlet"),
-                "right": (pybamm.Scalar(1), "Dirichlet"),
-            }
-        }
-        disc.bcs = boundary_conditions
-        disc.set_variable_slices([var])
-        grad_eqn_disc = disc.process_symbol(grad_eqn)
-        constant_y = np.ones((total_npts, 1))
-        np.testing.assert_array_equal(
-            grad_eqn_disc.evaluate(None, constant_y), np.zeros((total_npts_edges, 1))
-        )
-        # grad(r) == 1
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(0), "Dirichlet"),
-                "right": (pybamm.Scalar(1), "Dirichlet"),
-            }
-        }
-        disc.bcs = boundary_conditions
-        y_linear = np.tile(submesh.nodes, sec_npts)
-        grad_eqn_disc = disc.process_symbol(grad_eqn)
-        np.testing.assert_array_almost_equal(
-            grad_eqn_disc.evaluate(None, y_linear), np.ones((total_npts_edges, 1))
-        )
-        # div(grad r^2) = 6
-        N = pybamm.grad(var)
-        div_eqn = pybamm.div(N)
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(0), "Dirichlet"),
-                "right": (pybamm.Scalar(1), "Dirichlet"),
-            }
-        }
-        disc.bcs = boundary_conditions
-        div_eqn_disc = disc.process_symbol(div_eqn)
-        y_squared = np.tile(submesh.nodes ** 2, sec_npts)
-        div_eval = div_eqn_disc.evaluate(None, y_squared)
-        div_eval = np.reshape(div_eval, [sec_npts, npts])
-        np.testing.assert_array_almost_equal(
-            div_eval[:, :-1],
-            6 * np.ones([sec_npts, npts - 1]),
-        )
-
-    def test_grad_div_shapes_Neumann_bcs(self):
-        """Test grad and div with Neumann boundary conditions (applied by div on N)"""
-        whole_cell = ["negative electrode", "separator", "positive electrode"]
-
-        # create discretisation
-        mesh = get_mesh_for_testing()
-        spatial_methods = {"macroscale": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        combined_submesh = mesh.combine_submeshes(*whole_cell)
-
-        # grad
-        var = pybamm.Variable("var", domain=whole_cell)
-        grad_eqn = pybamm.grad(var)
-        disc.set_variable_slices([var])
-        grad_eqn_disc = disc.process_symbol(grad_eqn)
-
-        constant_y = np.ones_like(combined_submesh.nodes[:, np.newaxis])
-        np.testing.assert_array_equal(
-            grad_eqn_disc.evaluate(None, constant_y),
-            np.zeros_like(combined_submesh.edges[1:-1][:, np.newaxis]),
-        )
-
-        # div
-        N = pybamm.grad(var)
-        div_eqn = pybamm.div(N)
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(1), "Neumann"),
-                "right": (pybamm.Scalar(1), "Neumann"),
-            }
-        }
-        disc.bcs = boundary_conditions
-        div_eqn_disc = disc.process_symbol(div_eqn)
-
-        # Linear y should have laplacian zero
-        linear_y = combined_submesh.nodes
-        np.testing.assert_array_almost_equal(
-            grad_eqn_disc.evaluate(None, linear_y),
-            np.ones_like(combined_submesh.edges[1:-1][:, np.newaxis]),
-        )
-        np.testing.assert_array_almost_equal(
-            div_eqn_disc.evaluate(None, linear_y),
-            np.zeros_like(combined_submesh.nodes[:, np.newaxis]),
-        )
-
-    def test_grad_div_shapes_Dirichlet_and_Neumann_bcs(self):
-        """
-        Test grad and div with Dirichlet boundary conditions (applied by grad on c) on
-        one side and Neumann boundary conditions (applied by div on N) on the other
-        in Cartesian coordinates
-        """
-        whole_cell = ["negative electrode", "separator", "positive electrode"]
-
-        # create discretisation
-        mesh = get_mesh_for_testing()
-        spatial_methods = {"macroscale": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        combined_submesh = mesh.combine_submeshes(*whole_cell)
-
-        # grad
-        var = pybamm.Variable("var", domain=whole_cell)
-        grad_eqn = pybamm.grad(var)
-        disc.set_variable_slices([var])
-
-        # div
-        N = pybamm.grad(var)
-        div_eqn = pybamm.div(N)
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(1), "Dirichlet"),
-                "right": (pybamm.Scalar(0), "Neumann"),
-            }
-        }
-        disc.bcs = boundary_conditions
-        grad_eqn_disc = disc.process_symbol(grad_eqn)
-        div_eqn_disc = disc.process_symbol(div_eqn)
-
-        # Constant y should have gradient and laplacian zero
-        constant_y = np.ones_like(combined_submesh.nodes[:, np.newaxis])
-        np.testing.assert_array_equal(
-            grad_eqn_disc.evaluate(None, constant_y),
-            np.zeros_like(combined_submesh.edges[:, np.newaxis]),
-        )
-        np.testing.assert_array_almost_equal(
-            div_eqn_disc.evaluate(None, constant_y),
-            np.zeros_like(combined_submesh.nodes[:, np.newaxis]),
-        )
-
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(1), "Neumann"),
-                "right": (pybamm.Scalar(1), "Dirichlet"),
-            }
-        }
-        disc.bcs = boundary_conditions
-        grad_eqn_disc = disc.process_symbol(grad_eqn)
-        div_eqn_disc = disc.process_symbol(div_eqn)
-
-        # Linear y should have gradient one and laplacian zero
-        linear_y = combined_submesh.nodes
-        np.testing.assert_array_almost_equal(
-            grad_eqn_disc.evaluate(None, linear_y),
-            np.ones_like(combined_submesh.edges[:, np.newaxis]),
-        )
-        np.testing.assert_array_almost_equal(
-            div_eqn_disc.evaluate(None, linear_y),
-            np.zeros_like(combined_submesh.nodes[:, np.newaxis]),
-        )
-
-    def test_cylindrical_grad_div_shapes_Neumann_bcs(self):
-        """
-        Test grad and div with Neumann boundary conditions (applied by div on N)
-        in cylindrical polar coordinates
-        """
-
-        # create discretisation
-        mesh = get_cylindrical_mesh_for_testing()
-        spatial_methods = {"current collector": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        submesh = mesh["current collector"]
-        npts = submesh.npts
-        npts_edges = submesh.npts + 1
-
-        # grad
-        var = pybamm.Variable("var", domain="current collector")
-        grad_eqn = pybamm.grad(var)
-        disc.set_variable_slices([var])
-        grad_eqn_disc = disc.process_symbol(grad_eqn)
-
-        constant_y = np.ones((npts, 1))
-        np.testing.assert_array_equal(
-            grad_eqn_disc.evaluate(None, constant_y), np.zeros((npts_edges - 2, 1))
-        )
-
-        y_linear = submesh.nodes
-        np.testing.assert_array_almost_equal(
-            grad_eqn_disc.evaluate(None, y_linear), np.ones((npts_edges - 2, 1))
-        )
-
-        # div(grad(r^2)) = 4 , N_left = 2*r_inner, N_right = 2
-        N = pybamm.grad(var)
-        div_eqn = pybamm.div(N)
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(2 * submesh.edges[0]), "Neumann"),
-                "right": (pybamm.Scalar(2), "Neumann"),
-            }
-        }
-        disc.bcs = boundary_conditions
-        div_eqn_disc = disc.process_symbol(div_eqn)
-        y_squared = submesh.nodes ** 2
-        np.testing.assert_array_almost_equal(
-            div_eqn_disc.evaluate(None, y_squared),
-            4 * np.ones((npts, 1)),
-        )
-
-    def test_spherical_grad_div_shapes_Neumann_bcs(self):
-        """
-        Test grad and div with Neumann boundary conditions (applied by div on N)
-        in spherical polar coordinates
-        """
-
-        # create discretisation
-        mesh = get_mesh_for_testing()
-        spatial_methods = {"negative particle": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        combined_submesh = mesh.combine_submeshes("negative particle")
-
-        # grad
-        var = pybamm.Variable("var", domain="negative particle")
-        grad_eqn = pybamm.grad(var)
-        disc.set_variable_slices([var])
-        grad_eqn_disc = disc.process_symbol(grad_eqn)
-
-        constant_y = np.ones_like(combined_submesh.nodes[:, np.newaxis])
-        np.testing.assert_array_equal(
-            grad_eqn_disc.evaluate(None, constant_y),
-            np.zeros_like(combined_submesh.edges[1:-1][:, np.newaxis]),
-        )
-
-        linear_y = combined_submesh.nodes
-        np.testing.assert_array_almost_equal(
-            grad_eqn_disc.evaluate(None, linear_y),
-            np.ones_like(combined_submesh.edges[1:-1][:, np.newaxis]),
-        )
-
-        # div(grad(r^2)) = 6 , N_left = 0, N_right = 2
-        N = pybamm.grad(var)
-        div_eqn = pybamm.div(N)
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(0), "Neumann"),
-                "right": (pybamm.Scalar(2), "Neumann"),
-            }
-        }
-        disc.bcs = boundary_conditions
-        div_eqn_disc = disc.process_symbol(div_eqn)
-        quadratic_y = combined_submesh.nodes ** 2
-        np.testing.assert_array_almost_equal(
-            div_eqn_disc.evaluate(None, quadratic_y),
-            6 * np.ones((combined_submesh.npts, 1)),
-        )
-
-    def test_p2d_spherical_grad_div_shapes_Neumann_bcs(self):
-        """
-        Test grad and div with Dirichlet boundary conditions (applied by grad on var)
-        in the pseudo 2-dimensional case
-        """
-
-        mesh = get_p2d_mesh_for_testing()
-        spatial_methods = {"negative particle": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-        n_mesh = mesh["negative particle"]
-        mesh.add_ghost_meshes()
-        disc.mesh.add_ghost_meshes()
-        prim_pts = n_mesh.npts
-        sec_pts = mesh["negative electrode"].npts
-
-        # grad
-        var = pybamm.Variable(
-            "var",
-            domain=["negative particle"],
-            auxiliary_domains={"secondary": "negative electrode"},
-        )
-        grad_eqn = pybamm.grad(var)
-        disc.set_variable_slices([var])
-        grad_eqn_disc = disc.process_symbol(grad_eqn)
-        constant_y = np.kron(np.ones(sec_pts), np.ones(prim_pts))
-        grad_eval = grad_eqn_disc.evaluate(None, constant_y)
-        grad_eval = np.reshape(grad_eval, [sec_pts, prim_pts - 1])
-        np.testing.assert_array_equal(grad_eval, np.zeros([sec_pts, prim_pts - 1]))
-
-        # div(grad r^2) = 6, N_left = 0, N_right = 2
-        N = pybamm.grad(var)
-        div_eqn = pybamm.div(N)
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(0), "Neumann"),
-                "right": (pybamm.Scalar(2), "Neumann"),
-            }
-        }
-        disc.bcs = boundary_conditions
-        div_eqn_disc = disc.process_symbol(div_eqn)
-        y_squared = np.tile(n_mesh.nodes ** 2, sec_pts)
-        div_eval = div_eqn_disc.evaluate(None, y_squared)
-        div_eval = np.reshape(div_eval, [sec_pts, prim_pts])
-        np.testing.assert_array_almost_equal(div_eval, 6 * np.ones([sec_pts, prim_pts]))
-
-    def test_grad_div_shapes_mixed_domain(self):
-        """
-        Test grad and div with Dirichlet boundary conditions (applied by grad on var)
-        """
-        # create discretisation
-        mesh = get_mesh_for_testing()
-        spatial_methods = {"macroscale": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        # grad
-        var = pybamm.Variable("var", domain=["negative electrode", "separator"])
-        grad_eqn = pybamm.grad(var)
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(1), "Dirichlet"),
-                "right": (pybamm.Scalar(1), "Dirichlet"),
-            }
-        }
-        disc.bcs = boundary_conditions
-
-        disc.set_variable_slices([var])
-
-        grad_eqn_disc = disc.process_symbol(grad_eqn)
-
-        combined_submesh = mesh.combine_submeshes("negative electrode", "separator")
-        constant_y = np.ones_like(combined_submesh.nodes[:, np.newaxis])
-        np.testing.assert_array_equal(
-            grad_eqn_disc.evaluate(None, constant_y),
-            np.zeros_like(combined_submesh.edges[:, np.newaxis]),
-        )
-
-        # div: test on linear y (should have laplacian zero) so change bcs
-        linear_y = combined_submesh.nodes
-        N = pybamm.grad(var)
-        div_eqn = pybamm.div(N)
-        boundary_conditions = {
-            var.id: {
-                "left": (pybamm.Scalar(0), "Dirichlet"),
-                "right": (pybamm.Scalar(combined_submesh.edges[-1]), "Dirichlet"),
-            }
-        }
-        disc.bcs = boundary_conditions
-
-        grad_eqn_disc = disc.process_symbol(grad_eqn)
-        np.testing.assert_array_almost_equal(
-            grad_eqn_disc.evaluate(None, linear_y),
-            np.ones_like(combined_submesh.edges[:, np.newaxis]),
-        )
-
-        div_eqn_disc = disc.process_symbol(div_eqn)
-        np.testing.assert_array_almost_equal(
-            div_eqn_disc.evaluate(None, linear_y),
-            np.zeros_like(combined_submesh.nodes[:, np.newaxis]),
-        )
-
-    def test_definite_integral(self):
-        # create discretisation
-        mesh = get_mesh_for_testing(xpts=200, rpts=200)
-        spatial_methods = {
-            "macroscale": pybamm.FiniteVolume(),
-            "negative particle": pybamm.FiniteVolume(),
-            "positive particle": pybamm.FiniteVolume(),
-            "current collector": pybamm.ZeroDimensionalSpatialMethod(),
-        }
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-        # lengths
-        ln = mesh["negative electrode"].edges[-1]
-        ls = mesh["separator"].edges[-1] - ln
-        lp = 1 - (ln + ls)
-
-        # macroscale variable
-        var = pybamm.Variable("var", domain=["negative electrode", "separator"])
-        x = pybamm.SpatialVariable("x", ["negative electrode", "separator"])
-        integral_eqn = pybamm.Integral(var, x)
-        disc.set_variable_slices([var])
-        integral_eqn_disc = disc.process_symbol(integral_eqn)
-
-        combined_submesh = mesh.combine_submeshes("negative electrode", "separator")
-        constant_y = np.ones_like(combined_submesh.nodes[:, np.newaxis])
-        self.assertEqual(integral_eqn_disc.evaluate(None, constant_y), ln + ls)
-        linear_y = combined_submesh.nodes
-        np.testing.assert_array_almost_equal(
-            integral_eqn_disc.evaluate(None, linear_y), (ln + ls) ** 2 / 2
-        )
-        cos_y = np.cos(combined_submesh.nodes[:, np.newaxis])
-        np.testing.assert_array_almost_equal(
-            integral_eqn_disc.evaluate(None, cos_y), np.sin(ln + ls), decimal=4
-        )
-
-        # domain not starting at zero
-        var = pybamm.Variable("var", domain=["separator", "positive electrode"])
-        x = pybamm.SpatialVariable("x", ["separator", "positive electrode"])
-        integral_eqn = pybamm.Integral(var, x)
-        disc.set_variable_slices([var])
-        integral_eqn_disc = disc.process_symbol(integral_eqn)
-
-        combined_submesh = mesh.combine_submeshes("separator", "positive electrode")
-        constant_y = np.ones_like(combined_submesh.nodes[:, np.newaxis])
-        self.assertEqual(integral_eqn_disc.evaluate(None, constant_y), ls + lp)
-        linear_y = combined_submesh.nodes
-        self.assertAlmostEqual(
-            integral_eqn_disc.evaluate(None, linear_y)[0][0], (1 - (ln) ** 2) / 2
-        )
-        cos_y = np.cos(combined_submesh.nodes[:, np.newaxis])
-        np.testing.assert_array_almost_equal(
-            integral_eqn_disc.evaluate(None, cos_y), np.sin(1) - np.sin(ln), decimal=4
-        )
-
-        # microscale variable
-        var = pybamm.Variable("var", domain=["negative particle"])
-        r = pybamm.SpatialVariable("r", ["negative particle"])
-        integral_eqn = pybamm.Integral(var, r)
-        disc.set_variable_slices([var])
-        integral_eqn_disc = disc.process_symbol(integral_eqn)
-
-        constant_y = np.ones_like(mesh["negative particle"].nodes[:, np.newaxis])
-        np.testing.assert_array_almost_equal(
-            integral_eqn_disc.evaluate(None, constant_y), 4 * np.pi / 3
-        )
-        linear_y = mesh["negative particle"].nodes
-        np.testing.assert_array_almost_equal(
-            integral_eqn_disc.evaluate(None, linear_y), np.pi, decimal=4
-        )
-        one_over_y = 1 / mesh["negative particle"].nodes
-        np.testing.assert_array_almost_equal(
-            integral_eqn_disc.evaluate(None, one_over_y), 2 * np.pi, decimal=3
-        )
-
-        # cylindrical coordinates
-        mesh = get_cylindrical_mesh_for_testing(rcellpts=200)
-        spatial_methods = {"current collector": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        var = pybamm.Variable("var", domain=["current collector"])
-        r = pybamm.SpatialVariable("r", ["current collector"])
-        integral_eqn = pybamm.Integral(var, r)
-        disc.set_variable_slices([var])
-        integral_eqn_disc = disc.process_symbol(integral_eqn)
-
-        pts = mesh["current collector"].nodes
-        r0 = mesh["current collector"].edges[0]
-        constant_y = np.ones_like(pts[:, np.newaxis])
-        np.testing.assert_array_almost_equal(
-            integral_eqn_disc.evaluate(None, constant_y), np.pi * (1 - r0 ** 2)
-        )
-        linear_y = pts
-        np.testing.assert_array_almost_equal(
-            integral_eqn_disc.evaluate(None, linear_y),
-            2 * np.pi / 3 * (1 - r0 ** 3),
-            decimal=4,
-        )
-        one_over_y = 1 / pts
-        np.testing.assert_array_almost_equal(
-            integral_eqn_disc.evaluate(None, one_over_y),
-            2 * np.pi * (1 - r0),
-            decimal=3,
-        )
-
-        # test failure for secondary dimension column form
-        finite_volume = pybamm.FiniteVolume()
-        finite_volume.build(mesh)
-        with self.assertRaisesRegex(
-            NotImplementedError,
-            "Integral in secondary vector only implemented in 'row' form",
-        ):
-            finite_volume.definite_integral_matrix(var, "column", "secondary")
-
-    def test_integral_secondary_domain(self):
-        # create discretisation
-        mesh = get_1p1d_mesh_for_testing()
-        spatial_methods = {
-            "macroscale": pybamm.FiniteVolume(),
-            "negative particle": pybamm.FiniteVolume(),
-            "positive particle": pybamm.FiniteVolume(),
-            "current collector": pybamm.FiniteVolume(),
-        }
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-        # lengths
-        ln = mesh["negative electrode"].edges[-1]
-        ls = mesh["separator"].edges[-1] - ln
-        lp = 1 - (ln + ls)
-
-        var = pybamm.Variable(
-            "var",
-            domain="positive particle",
-            auxiliary_domains={
-                "secondary": "positive electrode",
-                "tertiary": "current collector",
-            },
-        )
-        x = pybamm.SpatialVariable("x", "positive electrode")
-        integral_eqn = pybamm.Integral(var, x)
-        disc.set_variable_slices([var])
-        integral_eqn_disc = disc.process_symbol(integral_eqn)
-
-        submesh = mesh["positive particle"]
-        constant_y = np.ones(
-            (
-                submesh.npts
-                * mesh["positive electrode"].npts
-                * mesh["current collector"].npts,
-                1,
-            )
-        )
-        np.testing.assert_array_almost_equal(
-            integral_eqn_disc.evaluate(None, constant_y),
-            lp * np.ones((submesh.npts * mesh["current collector"].npts, 1)),
-        )
-        linear_in_x = np.tile(
-            np.repeat(mesh["positive electrode"].nodes, submesh.npts),
-            mesh["current collector"].npts,
-        )
-        np.testing.assert_array_almost_equal(
-            integral_eqn_disc.evaluate(None, linear_in_x),
-            (1 - (ln + ls) ** 2)
-            / 2
-            * np.ones((submesh.npts * mesh["current collector"].npts, 1)),
-        )
-        linear_in_r = np.tile(
-            submesh.nodes,
-            mesh["positive electrode"].npts * mesh["current collector"].npts,
-        )
-        np.testing.assert_array_almost_equal(
-            integral_eqn_disc.evaluate(None, linear_in_r).flatten(),
-            lp * np.tile(submesh.nodes, mesh["current collector"].npts),
-        )
-        cos_y = np.cos(linear_in_x)
-        np.testing.assert_array_almost_equal(
-            integral_eqn_disc.evaluate(None, cos_y),
-            (np.sin(1) - np.sin(ln + ls))
-            * np.ones((submesh.npts * mesh["current collector"].npts, 1)),
-            decimal=4,
-        )
-
-    def test_integral_primary_then_secondary_same_result(self):
-        # Test that integrating in r then in x gives the same result as integrating in
-        # x then in r
-        # create discretisation
-        mesh = get_1p1d_mesh_for_testing()
-        spatial_methods = {
-            "macroscale": pybamm.FiniteVolume(),
-            "negative particle": pybamm.FiniteVolume(),
-            "positive particle": pybamm.FiniteVolume(),
-            "current collector": pybamm.FiniteVolume(),
-        }
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        var = pybamm.Variable(
-            "var",
-            domain="positive particle",
-            auxiliary_domains={
-                "secondary": "positive electrode",
-                "tertiary": "current collector",
-            },
-        )
-        x = pybamm.SpatialVariable("x", "positive electrode")
-        r = pybamm.SpatialVariable("r", "positive particle")
-        integral_eqn_x_then_r = pybamm.Integral(pybamm.Integral(var, x), r)
-        integral_eqn_r_then_x = pybamm.Integral(pybamm.Integral(var, r), x)
-
-        # discretise
-        disc.set_variable_slices([var])
-        integral_eqn_x_then_r_disc = disc.process_symbol(integral_eqn_x_then_r)
-        integral_eqn_r_then_x_disc = disc.process_symbol(integral_eqn_r_then_x)
-
-        # test
-        submesh = mesh["positive particle"]
-        cos_y = np.cos(
-            np.tile(
-                submesh.nodes,
-                mesh["positive electrode"].npts * mesh["current collector"].npts,
-            )
-        )
-        np.testing.assert_array_almost_equal(
-            integral_eqn_x_then_r_disc.evaluate(None, cos_y),
-            integral_eqn_r_then_x_disc.evaluate(None, cos_y),
-            decimal=4,
-        )
-
-    def test_integral_secondary_domain_on_edges_in_primary_domain(self):
-        # create discretisation
-        mesh = get_1p1d_mesh_for_testing()
-        spatial_methods = {
-            "macroscale": pybamm.FiniteVolume(),
-            "negative particle": pybamm.FiniteVolume(),
-            "positive particle": pybamm.FiniteVolume(),
-            "current collector": pybamm.FiniteVolume(),
-        }
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-        # lengths
-        ln = mesh["negative electrode"].edges[-1]
-        ls = mesh["separator"].edges[-1] - ln
-        lp = 1 - (ln + ls)
-
-        r_edge = pybamm.SpatialVariableEdge(
-            "r_p",
-            domain="positive particle",
-            auxiliary_domains={
-                "secondary": "positive electrode",
-                "tertiary": "current collector",
-            },
-        )
-
-        x = pybamm.SpatialVariable("x", "positive electrode")
-        integral_eqn = pybamm.Integral(r_edge, x)
-        integral_eqn_disc = disc.process_symbol(integral_eqn)
-
-        submesh = mesh["positive particle"]
-        np.testing.assert_array_almost_equal(
-            integral_eqn_disc.evaluate().flatten(),
-            lp
-            * np.tile(
-                np.linspace(0, 1, submesh.npts + 1), mesh["current collector"].npts
-            ),
-        )
-
-    def test_definite_integral_vector(self):
-        mesh = get_mesh_for_testing()
-        spatial_methods = {
-            "macroscale": pybamm.FiniteVolume(),
-            "negative particle": pybamm.FiniteVolume(),
-            "positive particle": pybamm.FiniteVolume(),
-        }
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-        var = pybamm.Variable("var", domain="negative electrode")
-        disc.set_variable_slices([var])
-
-        # row (default)
-        vec = pybamm.DefiniteIntegralVector(var)
-        vec_disc = disc.process_symbol(vec)
-        self.assertEqual(vec_disc.shape[0], 1)
-        self.assertEqual(vec_disc.shape[1], mesh["negative electrode"].npts)
-
-        # column
-        vec = pybamm.DefiniteIntegralVector(var, vector_type="column")
-        vec_disc = disc.process_symbol(vec)
-        self.assertEqual(vec_disc.shape[0], mesh["negative electrode"].npts)
-        self.assertEqual(vec_disc.shape[1], 1)
-
-    def test_indefinite_integral(self):
-
-        # create discretisation
-        mesh = get_mesh_for_testing()
-        spatial_methods = {
-            "macroscale": pybamm.FiniteVolume(),
-            "negative particle": pybamm.FiniteVolume(),
-            "positive particle": pybamm.FiniteVolume(),
-            "current collector": pybamm.ZeroDimensionalSpatialMethod(),
-        }
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        # input a phi, take grad, then integrate to recover phi approximation
-        # (need to test this way as check evaluated on edges using if has grad
-        # and no div)
-        phi = pybamm.Variable("phi", domain=["negative electrode", "separator"])
-        i = pybamm.grad(phi)  # create test current (variable on edges)
-
-        x = pybamm.SpatialVariable("x", ["negative electrode", "separator"])
-        int_grad_phi = pybamm.IndefiniteIntegral(i, x)
-        disc.set_variable_slices([phi])  # i is not a fundamental variable
-        # Set boundary conditions (required for shape but don't matter)
-        disc._bcs = {
-            phi.id: {
-                "left": (pybamm.Scalar(0), "Neumann"),
-                "right": (pybamm.Scalar(0), "Neumann"),
-            }
-        }
-        int_grad_phi_disc = disc.process_symbol(int_grad_phi)
-        left_boundary_value = pybamm.BoundaryValue(int_grad_phi, "left")
-        left_boundary_value_disc = disc.process_symbol(left_boundary_value)
-
-        combined_submesh = mesh.combine_submeshes("negative electrode", "separator")
-
-        # constant case
-        phi_exact = np.ones((combined_submesh.npts, 1))
-        phi_approx = int_grad_phi_disc.evaluate(None, phi_exact)
-        phi_approx += 1  # add constant of integration
-        np.testing.assert_array_equal(phi_exact, phi_approx)
-        self.assertEqual(left_boundary_value_disc.evaluate(y=phi_exact), 0)
-        # linear case
-        phi_exact = combined_submesh.nodes[:, np.newaxis]
-        phi_approx = int_grad_phi_disc.evaluate(None, phi_exact)
-        np.testing.assert_array_almost_equal(phi_exact, phi_approx)
-        self.assertEqual(left_boundary_value_disc.evaluate(y=phi_exact), 0)
-
-        # sine case
-        phi_exact = np.sin(combined_submesh.nodes[:, np.newaxis])
-        phi_approx = int_grad_phi_disc.evaluate(None, phi_exact)
-        np.testing.assert_array_almost_equal(phi_exact, phi_approx)
-        self.assertEqual(left_boundary_value_disc.evaluate(y=phi_exact), 0)
-
-        # --------------------------------------------------------------------
-        # region which doesn't start at zero
-        phi = pybamm.Variable("phi", domain=["separator", "positive electrode"])
-        i = pybamm.grad(phi)  # create test current (variable on edges)
-        x = pybamm.SpatialVariable("x", ["separator", "positive electrode"])
-        int_grad_phi = pybamm.IndefiniteIntegral(i, x)
-        disc.set_variable_slices([phi])  # i is not a fundamental variable
-        disc._bcs = {
-            phi.id: {
-                "left": (pybamm.Scalar(0), "Neumann"),
-                "right": (pybamm.Scalar(0), "Neumann"),
-            }
-        }
-        int_grad_phi_disc = disc.process_symbol(int_grad_phi)
-        left_boundary_value = pybamm.BoundaryValue(int_grad_phi, "left")
-        left_boundary_value_disc = disc.process_symbol(left_boundary_value)
-        combined_submesh = mesh.combine_submeshes("separator", "positive electrode")
-
-        # constant case
-        phi_exact = np.ones((combined_submesh.npts, 1))
-        phi_approx = int_grad_phi_disc.evaluate(None, phi_exact)
-        phi_approx += 1  # add constant of integration
-        np.testing.assert_array_equal(phi_exact, phi_approx)
-        self.assertEqual(left_boundary_value_disc.evaluate(y=phi_exact), 0)
-
-        # linear case
-        phi_exact = combined_submesh.nodes[:, np.newaxis] - combined_submesh.edges[0]
-        phi_approx = int_grad_phi_disc.evaluate(None, phi_exact)
-        np.testing.assert_array_almost_equal(phi_exact, phi_approx)
-        np.testing.assert_array_almost_equal(
-            left_boundary_value_disc.evaluate(y=phi_exact), 0
-        )
-
-        # sine case
-        phi_exact = np.sin(
-            combined_submesh.nodes[:, np.newaxis] - combined_submesh.edges[0]
-        )
-        phi_approx = int_grad_phi_disc.evaluate(None, phi_exact)
-        np.testing.assert_array_almost_equal(phi_exact, phi_approx)
-        np.testing.assert_array_almost_equal(
-            left_boundary_value_disc.evaluate(y=phi_exact), 0
-        )
-
-        # --------------------------------------------------------------------
-        # indefinite integral of a spatial variable
-        x = pybamm.SpatialVariable("x", ["negative electrode", "separator"])
-        x_edge = pybamm.SpatialVariableEdge("x", ["negative electrode", "separator"])
-        int_x = pybamm.IndefiniteIntegral(x, x)
-        int_x_edge = pybamm.IndefiniteIntegral(x_edge, x)
-
-        x_disc = disc.process_symbol(x)
-        x_edge_disc = disc.process_symbol(x_edge)
-        int_x_disc = disc.process_symbol(int_x)
-        int_x_edge_disc = disc.process_symbol(int_x_edge)
-
-        np.testing.assert_almost_equal(
-            int_x_disc.evaluate(), x_edge_disc.evaluate() ** 2 / 2
-        )
-        np.testing.assert_almost_equal(
-            int_x_edge_disc.evaluate(), x_disc.evaluate() ** 2 / 2, decimal=4
-        )
-
-        # --------------------------------------------------------------------
-        # micrsoscale case
-        c = pybamm.Variable("c", domain=["negative particle"])
-        N = pybamm.grad(c)  # create test flux (variable on edges)
-        r_n = pybamm.SpatialVariable("r_n", ["negative particle"])
-        c_integral = pybamm.IndefiniteIntegral(N, r_n)
-        disc.set_variable_slices([c])  # N is not a fundamental variable
-        disc._bcs = {
-            c.id: {
-                "left": (pybamm.Scalar(0), "Neumann"),
-                "right": (pybamm.Scalar(0), "Neumann"),
-            }
-        }
-
-        c_integral_disc = disc.process_symbol(c_integral)
-        left_boundary_value = pybamm.BoundaryValue(c_integral, "left")
-        left_boundary_value_disc = disc.process_symbol(left_boundary_value)
-        combined_submesh = mesh["negative particle"]
-
-        # constant case
-        c_exact = np.ones((combined_submesh.npts, 1))
-        c_approx = c_integral_disc.evaluate(None, c_exact)
-        c_approx += 1  # add constant of integration
-        np.testing.assert_array_equal(c_exact, c_approx)
-        self.assertEqual(left_boundary_value_disc.evaluate(y=c_exact), 0)
-
-        # linear case
-        c_exact = combined_submesh.nodes[:, np.newaxis]
-        c_approx = c_integral_disc.evaluate(None, c_exact)
-        np.testing.assert_array_almost_equal(c_exact, c_approx)
-        np.testing.assert_array_almost_equal(
-            left_boundary_value_disc.evaluate(y=c_exact), 0
-        )
-
-        # sine case
-        c_exact = np.sin(combined_submesh.nodes[:, np.newaxis])
-        c_approx = c_integral_disc.evaluate(None, c_exact)
-        np.testing.assert_array_almost_equal(c_exact, c_approx, decimal=3)
-        np.testing.assert_array_almost_equal(
-            left_boundary_value_disc.evaluate(y=c_exact), 0
-        )
-
-    def test_backward_indefinite_integral(self):
-
-        # create discretisation
-        mesh = get_mesh_for_testing()
-        spatial_methods = {"macroscale": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        # --------------------------------------------------------------------
-        # region which doesn't start at zero
-        phi = pybamm.Variable("phi", domain=["negative electrode", "separator"])
-        i = pybamm.grad(phi)  # create test current (variable on edges)
-        x = pybamm.SpatialVariable("x", ["negative electrode", "separator"])
-        int_grad_phi = pybamm.BackwardIndefiniteIntegral(i, x)
-        disc.set_variable_slices([phi])  # i is not a fundamental variable
-        disc._bcs = {
-            phi.id: {
-                "left": (pybamm.Scalar(0), "Neumann"),
-                "right": (pybamm.Scalar(0), "Neumann"),
-            }
-        }
-        int_grad_phi_disc = disc.process_symbol(int_grad_phi)
-        right_boundary_value = pybamm.BoundaryValue(int_grad_phi, "right")
-        right_boundary_value_disc = disc.process_symbol(right_boundary_value)
-        combined_submesh = mesh.combine_submeshes("negative electrode", "separator")
-
-        # Test that the backward_integral(grad(phi)) = -phi
-        # constant case
-        phi_exact = np.ones((combined_submesh.npts, 1))
-        phi_approx = int_grad_phi_disc.evaluate(None, phi_exact)
-        phi_approx += 1  # add constant of integration
-        np.testing.assert_array_equal(phi_exact, phi_approx)
-        self.assertEqual(right_boundary_value_disc.evaluate(y=phi_exact), 0)
-
-        # linear case
-        phi_exact = combined_submesh.nodes - combined_submesh.edges[-1]
-        phi_approx = int_grad_phi_disc.evaluate(None, phi_exact).flatten()
-        np.testing.assert_array_almost_equal(phi_exact, -phi_approx)
-        np.testing.assert_array_almost_equal(
-            right_boundary_value_disc.evaluate(y=phi_exact), 0
-        )
-
-        # sine case
-        phi_exact = np.sin(combined_submesh.nodes - combined_submesh.edges[-1])
-        phi_approx = int_grad_phi_disc.evaluate(None, phi_exact).flatten()
-        np.testing.assert_array_almost_equal(phi_exact, -phi_approx)
-        np.testing.assert_array_almost_equal(
-            right_boundary_value_disc.evaluate(y=phi_exact), 0
-        )
-
-    def test_indefinite_integral_of_broadcasted_to_cell_edges(self):
-        # create discretisation
-        mesh = get_mesh_for_testing()
-        spatial_methods = {
-            "macroscale": pybamm.FiniteVolume(),
-            "negative particle": pybamm.FiniteVolume(),
-            "positive particle": pybamm.FiniteVolume(),
-            "current collector": pybamm.ZeroDimensionalSpatialMethod(),
-        }
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        # make a variable 'phi' and a vector 'i' which is broadcast onto edges
-        # the integral of this should then be put onto the nodes
-        phi = pybamm.Variable("phi", domain=["negative electrode", "separator"])
-        i = pybamm.PrimaryBroadcastToEdges(1, phi.domain)
-        x = pybamm.SpatialVariable("x", phi.domain)
-        disc.set_variable_slices([phi])
-        combined_submesh = mesh.combine_submeshes("negative electrode", "separator")
-        x_end = combined_submesh.edges[-1]
-
-        # take indefinite integral
-        int_phi = pybamm.IndefiniteIntegral(i * phi, x)
-        # take integral again
-        int_int_phi = pybamm.Integral(int_phi, x)
-        int_int_phi_disc = disc.process_symbol(int_int_phi)
-
-        # constant case
-        phi_exact = np.ones_like(combined_submesh.nodes)
-        phi_approx = int_int_phi_disc.evaluate(None, phi_exact)
-        np.testing.assert_array_almost_equal(x_end ** 2 / 2, phi_approx)
-
-        # linear case
-        phi_exact = combined_submesh.nodes[:, np.newaxis]
-        phi_approx = int_int_phi_disc.evaluate(None, phi_exact)
-        np.testing.assert_array_almost_equal(x_end ** 3 / 6, phi_approx, decimal=4)
-
-    def test_indefinite_integral_on_nodes(self):
-        mesh = get_mesh_for_testing()
-        spatial_methods = {"macroscale": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        phi = pybamm.Variable("phi", domain=["negative electrode", "separator"])
-        x = pybamm.SpatialVariable("x", ["negative electrode", "separator"])
-
-        int_phi = pybamm.IndefiniteIntegral(phi, x)
-        disc.set_variable_slices([phi])
-        int_phi_disc = disc.process_symbol(int_phi)
-
-        combined_submesh = mesh.combine_submeshes("negative electrode", "separator")
-
-        # constant case
-        phi_exact = np.ones((combined_submesh.npts, 1))
-        int_phi_exact = combined_submesh.edges
-        int_phi_approx = int_phi_disc.evaluate(None, phi_exact).flatten()
-        np.testing.assert_array_equal(int_phi_exact, int_phi_approx)
-        # linear case
-        phi_exact = combined_submesh.nodes
-        int_phi_exact = combined_submesh.edges ** 2 / 2
-        int_phi_approx = int_phi_disc.evaluate(None, phi_exact).flatten()
-        np.testing.assert_array_almost_equal(int_phi_exact, int_phi_approx)
-        # cos case
-        phi_exact = np.cos(combined_submesh.nodes)
-        int_phi_exact = np.sin(combined_submesh.edges)
-        int_phi_approx = int_phi_disc.evaluate(None, phi_exact).flatten()
-        np.testing.assert_array_almost_equal(int_phi_exact, int_phi_approx, decimal=5)
-
-        # microscale case should fail
-        mesh = get_mesh_for_testing()
-        spatial_methods = {"negative particle": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        c = pybamm.Variable("c", domain=["negative particle"])
-        r = pybamm.SpatialVariable("r", ["negative particle"])
-
-        int_c = pybamm.IndefiniteIntegral(c, r)
-        disc.set_variable_slices([c])
-        with self.assertRaisesRegex(
-            NotImplementedError,
-            "Indefinite integral on a spherical polar domain is not implemented",
-        ):
-            disc.process_symbol(int_c)
-
-    def test_backward_indefinite_integral_on_nodes(self):
-        mesh = get_mesh_for_testing()
-        spatial_methods = {"macroscale": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        phi = pybamm.Variable("phi", domain=["negative electrode", "separator"])
-        x = pybamm.SpatialVariable("x", ["negative electrode", "separator"])
-
-        back_int_phi = pybamm.BackwardIndefiniteIntegral(phi, x)
-        disc.set_variable_slices([phi])
-        back_int_phi_disc = disc.process_symbol(back_int_phi)
-
-        combined_submesh = mesh.combine_submeshes("negative electrode", "separator")
-        edges = combined_submesh.edges
-
-        # constant case
-        phi_exact = np.ones((combined_submesh.npts, 1))
-        back_int_phi_exact = edges[-1] - edges
-        back_int_phi_approx = back_int_phi_disc.evaluate(None, phi_exact).flatten()
-        np.testing.assert_array_almost_equal(back_int_phi_exact, back_int_phi_approx)
-        # linear case
-        phi_exact = combined_submesh.nodes
-        back_int_phi_exact = edges[-1] ** 2 / 2 - edges ** 2 / 2
-        back_int_phi_approx = back_int_phi_disc.evaluate(None, phi_exact).flatten()
-        np.testing.assert_array_almost_equal(back_int_phi_exact, back_int_phi_approx)
-        # cos case
-        phi_exact = np.cos(combined_submesh.nodes)
-        back_int_phi_exact = np.sin(edges[-1]) - np.sin(edges)
-        back_int_phi_approx = back_int_phi_disc.evaluate(None, phi_exact).flatten()
-        np.testing.assert_array_almost_equal(
-            back_int_phi_exact, back_int_phi_approx, decimal=5
-        )
-
-    def test_forward_plus_backward_integral(self):
-        # Test that forward integral + backward integral = integral
-        mesh = get_mesh_for_testing()
-        spatial_methods = {"macroscale": pybamm.FiniteVolume()}
-        disc = pybamm.Discretisation(mesh, spatial_methods)
-
-        # On nodes
-        phi = pybamm.Variable("phi", domain=["separator", "positive electrode"])
-        x = pybamm.SpatialVariable("x", ["separator", "positive electrode"])
-
-        disc.set_variable_slices([phi])
-
-        full_int_phi = pybamm.PrimaryBroadcastToEdges(
-            pybamm.Integral(phi, x), ["separator", "positive electrode"]
-        )
-        full_int_phi_disc = disc.process_symbol(full_int_phi)
-        int_plus_back_int_phi = pybamm.IndefiniteIntegral(
-            phi, x
-        ) + pybamm.BackwardIndefiniteIntegral(phi, x)
-        int_plus_back_int_phi_disc = disc.process_symbol(int_plus_back_int_phi)
-
-        combined_submesh = mesh.combine_submeshes("separator", "positive electrode")
-
-        # test
-        for phi_exact in [
-            np.ones((combined_submesh.npts, 1)),
-            combined_submesh.nodes,
-            np.cos(combined_submesh.nodes),
-        ]:
-            np.testing.assert_array_almost_equal(
-                full_int_phi_disc.evaluate(y=phi_exact).flatten(),
-                int_plus_back_int_phi_disc.evaluate(y=phi_exact).flatten(),
-            )
-
     def test_discretise_spatial_variable(self):
-        # create discretisation
+        # Create discretisation
         mesh = get_mesh_for_testing()
         spatial_methods = {
             "macroscale": pybamm.FiniteVolume(),
@@ -1329,14 +152,14 @@ class TestFiniteVolume(unittest.TestCase):
         }
         disc = pybamm.Discretisation(mesh, spatial_methods)
 
-        # space
+        # macroscale
         x1 = pybamm.SpatialVariable("x", ["negative electrode"])
         x1_disc = disc.process_symbol(x1)
         self.assertIsInstance(x1_disc, pybamm.Vector)
         np.testing.assert_array_equal(
             x1_disc.evaluate(), disc.mesh["negative electrode"].nodes[:, np.newaxis]
         )
-
+        # macroscale with concatenation
         x2 = pybamm.SpatialVariable("x", ["negative electrode", "separator"])
         x2_disc = disc.process_symbol(x2)
         self.assertIsInstance(x2_disc, pybamm.Vector)
@@ -1346,7 +169,7 @@ class TestFiniteVolume(unittest.TestCase):
                 :, np.newaxis
             ],
         )
-
+        # microscale
         r = 3 * pybamm.SpatialVariable("r", ["negative particle"])
         r_disc = disc.process_symbol(r)
         self.assertIsInstance(r_disc, pybamm.Vector)
@@ -1355,10 +178,7 @@ class TestFiniteVolume(unittest.TestCase):
         )
 
     def test_mass_matrix_shape(self):
-        """
-        Test mass matrix shape
-        """
-        # one equation
+        # Create model
         whole_cell = ["negative electrode", "separator", "positive electrode"]
         c = pybamm.Variable("c", domain=whole_cell)
         N = pybamm.grad(c)
@@ -1370,22 +190,19 @@ class TestFiniteVolume(unittest.TestCase):
         }
         model.variables = {"c": c, "N": N}
 
-        # create discretisation
+        # Create discretisation
         mesh = get_mesh_for_testing()
         spatial_methods = {"macroscale": pybamm.FiniteVolume()}
         disc = pybamm.Discretisation(mesh, spatial_methods)
-
         combined_submesh = mesh.combine_submeshes(*whole_cell)
         disc.process_model(model)
 
-        # mass matrix
+        # Mass matrix
         mass = np.eye(combined_submesh.npts)
         np.testing.assert_array_equal(mass, model.mass_matrix.entries.toarray())
 
     def test_p2d_mass_matrix_shape(self):
-        """
-        Test mass matrix shape in the pseudo 2-dimensional case
-        """
+        # Create model
         c = pybamm.Variable(
             "c",
             domain=["negative particle"],
@@ -1399,11 +216,14 @@ class TestFiniteVolume(unittest.TestCase):
             c: {"left": (0, "Neumann"), "right": (0, "Dirichlet")}
         }
         model.variables = {"c": c, "N": N}
+
+        # Create discretisation
         mesh = get_p2d_mesh_for_testing()
         spatial_methods = {"negative particle": pybamm.FiniteVolume()}
         disc = pybamm.Discretisation(mesh, spatial_methods)
         disc.process_model(model)
 
+        # Mass matrix
         prim_pts = mesh["negative particle"].npts
         sec_pts = mesh["negative electrode"].npts
         mass_local = eye(prim_pts)
@@ -1413,15 +233,16 @@ class TestFiniteVolume(unittest.TestCase):
         )
 
     def test_jacobian(self):
+        # Create discretisation
         whole_cell = ["negative electrode", "separator", "positive electrode"]
-
-        # create discretisation
         mesh = get_mesh_for_testing()
         spatial_methods = {"macroscale": pybamm.FiniteVolume()}
         disc = pybamm.Discretisation(mesh, spatial_methods)
-
         combined_submesh = mesh.combine_submeshes(*whole_cell)
+        spatial_method = pybamm.FiniteVolume()
+        spatial_method.build(mesh)
 
+        # Setup variable
         var = pybamm.Variable("var", domain=whole_cell)
         disc.set_variable_slices([var])
         y = pybamm.StateVector(slice(0, combined_submesh.npts))
@@ -1438,8 +259,6 @@ class TestFiniteVolume(unittest.TestCase):
         eqn_disc = disc.process_symbol(eqn)
         eqn_jac = eqn_disc.jac(y)
         jacobian = eqn_jac.evaluate(y=y_test)
-        spatial_method = pybamm.FiniteVolume()
-        spatial_method.build(mesh)
         grad_matrix = spatial_method.gradient_matrix(whole_cell, {}).entries
         np.testing.assert_array_equal(jacobian.toarray()[1:-1], grad_matrix.toarray())
         np.testing.assert_array_equal(
@@ -1623,7 +442,7 @@ class TestFiniteVolume(unittest.TestCase):
             disc.process_symbol(downwind)
 
     def test_grad_div_with_bcs_on_tab(self):
-        # 2d macroscale
+        # Create discretisation
         mesh = get_1p1d_mesh_for_testing()
         spatial_methods = {
             "macroscale": pybamm.FiniteVolume(),
@@ -1632,9 +451,9 @@ class TestFiniteVolume(unittest.TestCase):
             "current collector": pybamm.FiniteVolume(),
         }
         disc = pybamm.Discretisation(mesh, spatial_methods)
-        y_test = np.ones(mesh["current collector"].npts)
 
         # var
+        y_test = np.ones(mesh["current collector"].npts)
         var = pybamm.Variable("var", domain="current collector")
         disc.set_variable_slices([var])
         # grad
@@ -1683,7 +502,7 @@ class TestFiniteVolume(unittest.TestCase):
         div_eqn_disc.evaluate(None, y_test)
 
     def test_neg_pos_bcs(self):
-        # 2d macroscale
+        # Create discretisation
         mesh = get_1p1d_mesh_for_testing()
         spatial_methods = {
             "macroscale": pybamm.FiniteVolume(),
